+++ lib/metro/stubs/api_service_stub.dart
@@ -4,7 +4,7 @@ String apiServiceStub(ReCase rc,
         {required ReCase model, required String baseUrl}) =>
     '''
 import 'package:flutter/material.dart';
-import '../../app/networking/dio/base_api_service.dart';
+import '/app/networking/dio/base_api_service.dart';
 ${baseUrl == "getEnv('API_BASE_URL')" ? "import 'package:nylo_framework/nylo_framework.dart';" : ""}
 ${model.originalText != 'Model' ? "import '../../app/models/${model.snakeCase}.dart';" : ""}
 
@@ -15,7 +15,7 @@ class ${rc.pascalCase}ApiService extends BaseApiService {
   String get baseUrl => $baseUrl;
 
 ${model.originalText != "Model" ? ''' 
-  /// Return a list of users
+  /// Return a list of ${model.pascalCase}
   Future<List<${model.pascalCase}>?> fetchAll({dynamic query}) async {
     return await network<List<${model.pascalCase}>>(
         request: (request) => request.get("/endpoint-path", queryParameters: query),