+++ lib/metro/stubs/widget_stateful_stub.dart
@@ -5,6 +5,9 @@ import 'package:flutter/material.dart';
 import 'package:nylo_framework/nylo_framework.dart';
 
 class ${rc.pascalCase} extends StatefulWidget {
+  
+  ${rc.pascalCase}({Key? key}) : super(key: key);
+
   @override
   _${rc.pascalCase}State createState() => _${rc.pascalCase}State();
 }
@@ -14,6 +17,7 @@ class _${rc.pascalCase}State extends NyState<${rc.pascalCase}> {
   @override
   init() async {
     super.init();
+    
   }
 
   @override